+++ upcloud/service/kubernetes_ctx_test.go
@@ -185,21 +185,6 @@ func TestKubernetesCtx(t *testing.T) {
 		})
 	})
 
-	t.Run("GetKubernetesPlans", func(t *testing.T) {
-		t.Parallel()
-
-		recordWithContext(t, "get_kubernetes_plans", func(ctx context.Context, t *testing.T, rec *recorder.Recorder, svc *Service, svcContext *ServiceContext) {
-			p, err := svcContext.GetKubernetesPlans(ctx, &request.GetKubernetesPlansRequest{})
-
-			require.NoError(t, err)
-			require.NotZero(t, p)
-
-			firstPlan := p[0]
-			require.NotZero(t, firstPlan.Description)
-			require.NotZero(t, firstPlan.Name)
-		})
-	})
-
 	t.Run("GetKubernetesVersions", func(t *testing.T) {
 		t.Parallel()